# index.js
import express from "express";
import bodyParser from "body-parser";
import swaggerUi from "swagger-ui-express";
import openapiSpecification from "./openapi.json" with { type: "json" };
const  app = express();
const PORT = 3000;
const tasks=[
{
id:1,
title:"learn",
done:true
},
{
id:2,
title:"build",
done:false
},
{
id:3,
title:"test",
done:true
}
];
app.use(bodyParser.json());
app.get("/", (req, res) => {
  console.log('[GET ROUTE]');
    res.json({name :"task API",version: "1.0",endpoints:["/tasks"]});

});
app.get("/health", (req,res)=>{
  res.json({status:"ok"});
});
app.get("/tasks" , (req,res)=>{
  res.json(tasks);
});
app.get("/tasks/:id", (req, res) => {
 const id=Number(req.params.id);
 const task=tasks.find(task=>task.id===id);
if (!task) {
  return res.status(404).json({
    error: `Task ${id} not found`
  });
}
res.json(task);
});
app.post("/tasks", (req,res)=>{
  const{title}=req.body;
  if(!title || title.trim()===""){
    return res.status(400).json({error:"Title is required"});
  }

const newTask={
  id: tasks.length +1,
  title: title.trim(),
  done: false
};
tasks.push(newTask);
res.status(201).json(newTask);
});
app.put("/tasks/:id", (req,res)=>{
const id=Number(req.params.id);
const task=tasks.find(task=>task.id===id);
if(!task){
  return res.status(404).json({error:`Task ${id} not found`});
}
const {title,done}=req.body;
if(Object.keys(req.body).length===0){
return res.status(400).json({
      error: "Request body cannot be empty"
    });
  }
if (title !== undefined) {

     if (typeof title !== "string" || title.trim() === "") {
      return res.status(400).json({
        error: "Title must be a non-empty string"
      });
    }

    task.title = title.trim();
  }
   if (done !== undefined) {
     if (typeof done !== "boolean") {
      return res.status(400).json({
        error: "Done must be true or false"
      });
    }
    task.done = done;
  }
   res.json(task);
});
app.delete("/tasks/:id", (req, res) => {
  const id = Number(req.params.id);
 const taskIndex = tasks.findIndex(task => task.id === id);
  if (taskIndex === -1) {
    return res.status(404).json({
      error: `Task ${id} not found`
    });
  }
  tasks.splice(taskIndex, 1);
  res.status(204).send();
});
app.use("/docs", swaggerUi.serve, swaggerUi.setup(openapiSpecification));
 app.listen(PORT, () => console.log(`Server running on port: http://localhost:${PORT}`));